+++ GRCFinancialControl.Core/Models/Reporting/PlannedVsActualData.cs
@@ -6,8 +6,8 @@ namespace GRCFinancialControl.Core.Models.Reporting
         public string EngagementDescription { get; set; } = string.Empty;
         public string PapdName { get; set; } = string.Empty;
         public string FiscalYear { get; set; } = string.Empty;
-        public double PlannedHours { get; set; }
-        public double ActualHours { get; set; }
-        public double Variance => PlannedHours - ActualHours;
+        public decimal PlannedHours { get; set; }
+        public decimal ActualHours { get; set; }
+        public decimal Variance => PlannedHours - ActualHours;
     }
 }
